Resupply orders and testing finished.
MaintenanceSupplies under pop was switched to an integer. I don't think
that fractional MSPs are consumed. Likewise fuel was switched to a float
from an int earlier, I do believe that fractional litres of fuel are
consumed.

+++ Pulsar4X/Pulsar4X.Lib/Entities/Population.cs
@@ -59,7 +59,7 @@ namespace Pulsar4X.Entities
         public float CivilianPopulation { get; set; }
         public float PopulationGrowthRate { get; set; }
         public float FuelStockpile { get; set; }
-        public float MaintenanceSupplies { get; set; }
+        public int MaintenanceSupplies { get; set; }
 
 
         public float PopulationWorkingInAgriAndEnviro
